+++ marketing/app/why/page.tsx
@@ -1,0 +1,26 @@
+export default function Page() {
+  return (
+    <article className="prose lg:prose-xl my-20">
+      <h1>Why Carbon</h1>
+      <p>
+        The founders of Carbon worked at various manufacturing companies over
+        the years and saw the same problems repeatedly.
+      </p>
+      <p>
+        As software became cloud-based, the software landlords started to
+        increase costs over time. You had to play by their rules. Every
+        customization would require working with a &apos;Software
+        Consultant&apos;
+      </p>
+      <p>
+        We realized it was time for change. Rather than charging outrageous
+        license and consultancy fees, we decided to make the software
+        open-source. Rather than just being a tenant, you now OWN the software.
+      </p>
+      <p>
+        And if you are a company that doesn&apos;t want to deal with the hosting
+        IT, we can help you get started with the deployment and support.
+      </p>
+    </article>
+  );
+}
