+++ marketing/tailwind.config.ts
@@ -84,6 +84,7 @@ const config = {
   },
   plugins: [
     require("tailwindcss-animate"),
+    require('@tailwindcss/typography'),
     function ({ matchUtilities, theme }: any) {
       matchUtilities(
         {
